Rose > Dev (#3)
* /factory/profile + register factory

* PUT /factories/:factory_id/profile

* profileInit

* Explore Page

* login /me

* create showcase PD

* insert showcase ID

* edit showcase

* create rfq MR + delete inspection_type

* rfq

* fix /explore, /bootstrap

* /factory/rfqs

* rfq-board

* rfq detail

* cronjob CC rfqs

* rfq detail

* หน้า แนะนำโรงงาน และ detail

* Edit boq

* rfq+ explore path

* rfq

* order detail

* regist

* wallet

* step0

* printShippingLabel

* step4

* notification

* CP order step 5

* rfqs.expired_date

* Favorites

* sse + chat

* chat read

* api factory-ideas showcases

* factoryRating

* boq chat

* ReadBufferSize

* api portal

* rfq api portal

* refactor code

* review

* delete dispute order

รอเฟสสองค่อยเปิด เพิ่มตาราง dispute แจ้งปัญหา order , order จะไม่ผิดถ้าไม่ปิดข้อพิพาธก่อน

---------

+++ tryly-server/internal/domain/review_models.go
@@ -7,7 +7,7 @@ type FactoryReview struct {
 	FactoryID      int64       `db:"factory_id" json:"factory_id"`
 	UserID         int64       `db:"user_id" json:"user_id"`
 	OrderID        *int64      `db:"order_id" json:"order_id,omitempty"`
-	Rating         int         `db:"rating" json:"rating"`
+	Rating         float64     `db:"rating" json:"rating"`
 	Comment        string      `db:"comment" json:"comment"`
 	ImageURLs      StringArray `db:"image_urls" json:"image_urls"`
 	CreatedAt      time.Time   `db:"created_at" json:"created_at"`
